requirements.txt
phidata
streamlit
duckduckgo-search
google-generativeai
python-dotenv
google-genai
youtube-transcript-api